requirements.txt
contourpy==1.2.1
cycler==0.12.1
distro==1.9.0
fonttools==4.53.0
joblib==1.4.2
JPype1==1.5.0
kiwisolver==1.4.5
matplotlib==3.9.0
numpy==2.0.0
packaging==24.1
pandas==2.2.2
pillow==10.3.0
pyparsing==3.1.2
pypdf==4.2.0
python-dateutil==2.9.0.post0
pytz==2024.1
scikit-learn==1.5.0
scipy==1.14.0
seaborn==0.13.2
six==1.16.0
tabula-py==2.9.3
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.1
